automation/requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-analytics-data>=0.18.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0  # Gemini API（AI要約用）

# Social Media Automation
tweepy>=4.14.0
feedparser>=6.0.10
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# AI要約（オプション）
openai>=1.0.0  # OpenAI GPT API（オプション）
anthropic>=0.18.0  # Claude API（オプション）
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0

# DMM Affiliate
dmm-search3>=2.0.0
